+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ---
 
 ## main
+* Artist browse
 
 ## 0.1.1 - 03/05/2023
 * Github actions for native image